Customer application is functional
The customer is now able to select which item ID they would like to
purchase. The logic then checks whether the quantity requested by the
customer is available. If so, the order is fulfilled and the product
quantity is updated in the database. If the customer has requested a
quantity that is not available, they are asked to modify their order
before it can be processed.